tidy: Re-enable check for inline unit tests
